+++ js/main.js
@@ -4,78 +4,9 @@
 //likes, число — количество лайков, поставленных фотографии. Случайное число от 15 до 200
 //comments, массив объектов — список комментариев, оставленных другими пользователями к этой фотографии. Количество комментариев к каждой фотографии — случайное число от 0 до 30.
 
-const NAME =[
-  'David',
-  'John',
-  'Sebastian',
-  'Franz',
-  'Wolfie',
-  'Ada'
-]
 
-const DESCRIPTION = [
-  'What a strike it was!',
-  'Awwesome summer!',
-  'Never imagine to be somewhere like here',
-  'Вот это да!',
-  'Невероятный закат',
-  'Все оттенки лета',
-  'Белые начинают и выигрывают'
-]
+import { similarPhotos } from './data';
 
-const MESSAGE = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-]
-
-
-function getRandomInteger (min, max) {
-  const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
-  const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
-  const result = Math.random() * (upper - lower + 1) + lower;
-
-  return Math.floor(result);
-}
-
-function createRandomIdFromRangeGenerator (min, max) {
-  const previousValues = [];
-
-  return function () {
-    let currentValue = getRandomInteger(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-      console.error('Перебраны все числа из диапазона от ' + min + ' до ' + max);
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-}
-
-
-const createPhoto = () => {
-  return {
-    id: getRandomInteger(1, 25),
-    url: 'photos/' + getRandomInteger(1, 25) + '.jpg',
-    description: DESCRIPTION[getRandomInteger(0, DESCRIPTION.length - 1)],
-    likes: getRandomInteger(15, 200),
-    comments: {
-      id: getRandomInteger(1, 250),
-      avatar: 'img/avatar-' + getRandomInteger(1, 6) + '.svg',
-      message: MESSAGE[getRandomInteger(0, MESSAGE.length - 1)],
-      name: NAME[getRandomInteger(0, NAME.length - 1)],
-    }
-  }
-}
-
-
-const similarPhotos = Array.from({length: 25}, createPhoto)
 
 console.log(similarPhotos);
